role

+++ src/controllers/aluno.controller.ts
@@ -4,6 +4,12 @@ import alunoService from '../services/aluno.service';
 class AlunoController {
   async create(req: Request, res: Response) {
     try {
+      if (!req.user) {
+        return res.status(401).json({ error: 'Acesso negado' });
+      }
+      
+      console.log("Criar aluno");
+
       const aluno = await alunoService.create(req.body);
       res.status(201).json(aluno);
     } catch (error: any) {
@@ -13,6 +19,8 @@ class AlunoController {
 
   async completeProfile(req: Request, res: Response) {
     try {
+      console.log("Completar o perfil do aluno");
+      
       const result = await alunoService.completeProfile(req.body)
       res.status(200).json(result)
     } catch (error: any) {
@@ -22,7 +30,18 @@ class AlunoController {
 
   async findAll(req: Request, res: Response) {
     try {
-      const alunos = await alunoService.findAll();
+
+      if (req.user && req.user.role === 'student') {
+        return res.status(403).json({ error: 'Acesso negado' });
+      }
+
+      if (!req.user) {
+        return res.status(401).json({ error: 'Acesso negado' });
+      }
+
+      console.log("Buscar todos");
+
+      const alunos = await alunoService.findAll(req.user);
       res.json(alunos);
     } catch (error: any) {
       res.status(500).json({ error: error.message });
@@ -31,6 +50,12 @@ class AlunoController {
 
   async findById(req: Request, res: Response) {
     try {
+      console.log("Buscar pelo Id");
+
+      if (!req.user) {
+        return res.status(401).json({ error: 'Acesso negado' });
+      }
+
       const aluno = await alunoService.findById(Number(req.params.id));
       if (aluno) {
         res.json(aluno);
@@ -44,6 +69,8 @@ class AlunoController {
 
   async update(req: Request, res: Response) {
     try {
+      console.log("Atualizar");
+
       const aluno = await alunoService.update(Number(req.params.id), req.body);
       res.json(aluno);
     } catch (error: any) {
@@ -53,6 +80,9 @@ class AlunoController {
 
   async delete(req: Request, res: Response) {
     try {
+      if (!req.user) {
+        return res.status(401).json({ error: 'Acesso negado' });
+      }
       await alunoService.delete(Number(req.params.id));
       res.status(204).send();
     } catch (error: any) {
@@ -62,6 +92,9 @@ class AlunoController {
 
   async findByTurma(req: Request, res: Response) {
     try {
+      if (!req.user) {
+        return res.status(401).json({ error: 'Acesso negado' });
+      }
       const alunos = await alunoService.findAlunosByTurma(Number(req.params.turmaId));
       res.json(alunos);
     } catch (error: any) {
