+++ src/controllers/disciplina.controller.ts
@@ -4,6 +4,9 @@ import disciplinaService from '../services/disciplina.service';
 class DisciplinaController {
   async create(req: Request, res: Response) {
     try {
+      if (!req.user) {
+        return res.status(401).json({ error: 'Acesso negado' });
+      }
       const disciplina = await disciplinaService.create(req.body);
       res.status(201).json(disciplina);
     } catch (error: any) {
@@ -13,6 +16,9 @@ class DisciplinaController {
 
   async findAll(req: Request, res: Response) {
     try {
+      if (!req.user) {
+        return res.status(401).json({ error: 'Acesso negado' });
+      }
       const disciplinas = await disciplinaService.findAll();
       res.json(disciplinas);
     } catch (error: any) {
@@ -21,6 +27,9 @@ class DisciplinaController {
   }
 
   async findById(req: Request, res: Response) {
+    if (!req.user) {
+      return res.status(401).json({ error: 'Acesso negado' });
+    }
     try {
       const disciplina = await disciplinaService.findById(Number(req.params.id));
       if (disciplina) {
@@ -35,6 +44,9 @@ class DisciplinaController {
 
   async update(req: Request, res: Response) {
     try {
+      if (!req.user || req.user.role !== 'admin') {
+        return res.status(401).json({ error: 'Acesso negado' });
+      }
       const disciplina = await disciplinaService.update(Number(req.params.id), req.body);
       res.json(disciplina);
     } catch (error: any) {
@@ -43,6 +55,9 @@ class DisciplinaController {
   }
 
   async delete(req: Request, res: Response) {
+    if(!req.user || req.user.role !== 'admin') {
+      return res.status(401).json({ error: 'Acesso negado' });
+    }
     try {
       await disciplinaService.delete(Number(req.params.id));
       res.status(204).send();
